implementa lógica para atualizar saldo do pagador/recebedor

+++ TaahjjimBank/src/main/java/handler/LambdaLiquidaAgendadasHandler.java
@@ -7,26 +7,32 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import model.TransacaoModel;
 import model.eStatusTransacao;
+import model.ContaBancariaModel;
 import service.DriverS3;
+import service.ContaBancariaService;
 
 import java.util.ArrayList;
 import java.util.List;
 
-// ess lambda será acionada automaticamente por mensagens da fila SQS
 public class LambdaLiquidaAgendadasHandler implements RequestHandler<SQSEvent, Void> {
 
-    private final ObjectMapper objectMapper = new ObjectMapper().registerModule(new JavaTimeModule());
-    private final DriverS3<TransacaoModel> driverS3 = new DriverS3<>("zupbankdatabase", TransacaoModel.class);
     private final String PATH_TRANSACOES = "dados/transacao/";
     private final String PATH_AGENDADAS = "dados/transacaoAgendada/";
+    private static final String BUCKET_NAME = "zupbankdatabase";
+
+    private final ObjectMapper objectMapper = new ObjectMapper().registerModule(new JavaTimeModule());
+    private final DriverS3<TransacaoModel> driverS3 = new DriverS3<>(BUCKET_NAME, TransacaoModel.class);
+
 
     @Override
     public Void handleRequest(SQSEvent event, Context context) {
+        ContaBancariaService contaService = new ContaBancariaService("zupbankdatabase", null);
+
         for (SQSEvent.SQSMessage message : event.getRecords()) {
             try {
                 TransacaoModel transacao = objectMapper.readValue(message.getBody(), TransacaoModel.class);
 
-                // autaliza status para CONCLUIDA e define dataTransacao atual
+                // atualiza status para CONCLUIDA e define dataTransacao atual
                 transacao.setStatusTransacao(eStatusTransacao.CONCLUIDA);
                 transacao.setDataTransacao(java.time.LocalDateTime.now());
 
@@ -50,12 +56,23 @@ public class LambdaLiquidaAgendadasHandler implements RequestHandler<SQSEvent, V
                 List<TransacaoModel> transacoesAgendadas = driverS3.readList(keyAgendada, TransacaoModel.class).orElse(new ArrayList<>());
                 transacoesAgendadas.removeIf(t -> t.getId().equals(transacao.getId()));
 
+                //atualiza saldo na origem
+                ContaBancariaModel contaOrigem = contaService.obter(transacao.getNumeroContaOrigem());
+                if (contaOrigem == null) throw new RuntimeException("Conta origem não encontrada");
+                contaOrigem.setSaldo(contaOrigem.getSaldo() - transacao.getValorTransacao());
+                contaService.salvar(contaOrigem);
+
+                //atualiza saldo no destino
+                ContaBancariaModel contaDestino = contaService.obter(transacao.getNumeroContaDestino());
+                if (contaDestino == null) throw new RuntimeException("Conta destino não encontrada");
+                contaDestino.setSaldo(contaDestino.getSaldo() + transacao.getValorTransacao());
+                contaService.salvar(contaOrigem);
+
                 if (transacoesAgendadas.isEmpty()) {
                     driverS3.deleteObject(keyAgendada);
                 } else {
                     driverS3.saveList(keyAgendada, transacoesAgendadas);
                 }
-
             } catch (Exception e) {
                 context.getLogger().log("Erro ao processar transação agendada: " + e.getMessage());
             }
